ServerApp/requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.17
django-cron==0.6.0
django-filter==24.3
djangorestframework==3.15.2
graphene==3.4.3
graphene-django==3.2.2
graphql-core==3.2.6
graphql-relay==3.2.0
kombu==5.4.2
promise==2.3
prompt_toolkit==3.0.48
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.2
vine==5.1.0
wcwidth==0.2.13
